Jabel, Generic injection and automatic code cleanup (#21)
* Enable jabel&generic injection, fix generic injection errors

* Enhanced for loops

* Add missing generics part 1

* Add missing generics part 2: getSubBlocks

* Add missing generics part 3: structure components

* Add missing generics part 4: addInformation

* MapGenTFHollowTree cleanup

* Replace an unreadable do-while loop with a stream filter-find

* Fix unnecessary boxing

* Fix unnecessary unboxing

* While->Enhanced for loop

* Use diamond operator where appropriate

* Collapse identical catch blocks

* String switch

* Use enhanced switch

* Fix compilation error after added generics

* Use instanceof pattern matching

* More missed generic parameters

* Use Java-style array declarations

* Unnecessary String conversions

* Unnecessary modifiers

* Unnecessary semicolons

* Unchecked warnings fixes

* Explicit array filling

* Manual min/max calculations

* Redundant vararg array creation

* Redundant type casts

* Unnecessary StringBuilder

* Remove redundant suppressions

* Apply spotless

* Fix ambiguous method call error

* Update baubles to fix runObfClient

+++ src/main/java/twilightforest/block/BlockTFCinderFurnace.java
@@ -130,17 +130,17 @@ public class BlockTFCinderFurnace extends BlockContainer {
             int l = random.nextInt(4) + 2;
 
             if (l == 4) {
-                world.spawnParticle("smoke", (double) (f - f3), (double) f1, (double) (f2 + f4), 0.0D, 0.0D, 0.0D);
-                world.spawnParticle("flame", (double) (f - f3), (double) f1, (double) (f2 + f4), 0.0D, 0.0D, 0.0D);
+                world.spawnParticle("smoke", f - f3, f1, f2 + f4, 0.0D, 0.0D, 0.0D);
+                world.spawnParticle("flame", f - f3, f1, f2 + f4, 0.0D, 0.0D, 0.0D);
             } else if (l == 5) {
-                world.spawnParticle("smoke", (double) (f + f3), (double) f1, (double) (f2 + f4), 0.0D, 0.0D, 0.0D);
-                world.spawnParticle("flame", (double) (f + f3), (double) f1, (double) (f2 + f4), 0.0D, 0.0D, 0.0D);
+                world.spawnParticle("smoke", f + f3, f1, f2 + f4, 0.0D, 0.0D, 0.0D);
+                world.spawnParticle("flame", f + f3, f1, f2 + f4, 0.0D, 0.0D, 0.0D);
             } else if (l == 2) {
-                world.spawnParticle("smoke", (double) (f + f4), (double) f1, (double) (f2 - f3), 0.0D, 0.0D, 0.0D);
-                world.spawnParticle("flame", (double) (f + f4), (double) f1, (double) (f2 - f3), 0.0D, 0.0D, 0.0D);
+                world.spawnParticle("smoke", f + f4, f1, f2 - f3, 0.0D, 0.0D, 0.0D);
+                world.spawnParticle("flame", f + f4, f1, f2 - f3, 0.0D, 0.0D, 0.0D);
             } else if (l == 3) {
-                world.spawnParticle("smoke", (double) (f + f4), (double) f1, (double) (f2 + f3), 0.0D, 0.0D, 0.0D);
-                world.spawnParticle("flame", (double) (f + f4), (double) f1, (double) (f2 + f3), 0.0D, 0.0D, 0.0D);
+                world.spawnParticle("smoke", f + f4, f1, f2 + f3, 0.0D, 0.0D, 0.0D);
+                world.spawnParticle("flame", f + f4, f1, f2 + f3, 0.0D, 0.0D, 0.0D);
             }
         }
     }
@@ -171,9 +171,9 @@ public class BlockTFCinderFurnace extends BlockContainer {
                             itemstack.stackSize -= j1;
                             EntityItem entityitem = new EntityItem(
                                     world,
-                                    (double) ((float) x + dx),
-                                    (double) ((float) y + dy),
-                                    (double) ((float) z + dz),
+                                    (float) x + dx,
+                                    (float) y + dy,
+                                    (float) z + dz,
                                     new ItemStack(itemstack.getItem(), j1, itemstack.getItemDamage()));
 
                             if (itemstack.hasTagCompound()) {
@@ -182,10 +182,9 @@ public class BlockTFCinderFurnace extends BlockContainer {
                             }
 
                             float pointFive = 0.05F;
-                            entityitem.motionX = (double) ((float) this.furnaceRandom.nextGaussian() * pointFive);
-                            entityitem.motionY = (double) ((float) this.furnaceRandom.nextGaussian() * pointFive
-                                    + 0.2F);
-                            entityitem.motionZ = (double) ((float) this.furnaceRandom.nextGaussian() * pointFive);
+                            entityitem.motionX = (float) this.furnaceRandom.nextGaussian() * pointFive;
+                            entityitem.motionY = (float) this.furnaceRandom.nextGaussian() * pointFive + 0.2F;
+                            entityitem.motionZ = (float) this.furnaceRandom.nextGaussian() * pointFive;
                             world.spawnEntityInWorld(entityitem);
                         }
                     }
