+++ src/main/java/twilightforest/client/particle/EntityTFBossTearFX.java
@@ -93,29 +93,29 @@ public class EntityTFBossTearFX extends EntityFX {
         float f14 = 1.0F;
         par1Tessellator.setColorOpaque_F(f14 * this.particleRed, f14 * this.particleGreen, f14 * this.particleBlue);
         par1Tessellator.addVertexWithUV(
-                (double) (f11 - par3 * f10 - par6 * f10),
-                (double) (f12 - par4 * f10),
-                (double) (f13 - par5 * f10 - par7 * f10),
-                (double) f6,
-                (double) f9);
+                f11 - par3 * f10 - par6 * f10,
+                f12 - par4 * f10,
+                f13 - par5 * f10 - par7 * f10,
+                f6,
+                f9);
         par1Tessellator.addVertexWithUV(
-                (double) (f11 - par3 * f10 + par6 * f10),
-                (double) (f12 + par4 * f10),
-                (double) (f13 - par5 * f10 + par7 * f10),
-                (double) f6,
-                (double) f8);
+                f11 - par3 * f10 + par6 * f10,
+                f12 + par4 * f10,
+                f13 - par5 * f10 + par7 * f10,
+                f6,
+                f8);
         par1Tessellator.addVertexWithUV(
-                (double) (f11 + par3 * f10 + par6 * f10),
-                (double) (f12 + par4 * f10),
-                (double) (f13 + par5 * f10 + par7 * f10),
-                (double) f7,
-                (double) f8);
+                f11 + par3 * f10 + par6 * f10,
+                f12 + par4 * f10,
+                f13 + par5 * f10 + par7 * f10,
+                f7,
+                f8);
         par1Tessellator.addVertexWithUV(
-                (double) (f11 + par3 * f10 - par6 * f10),
-                (double) (f12 - par4 * f10),
-                (double) (f13 + par5 * f10 - par7 * f10),
-                (double) f7,
-                (double) f9);
+                f11 + par3 * f10 - par6 * f10,
+                f12 - par4 * f10,
+                f13 + par5 * f10 - par7 * f10,
+                f7,
+                f9);
     }
 
 }
